Update driver: Exception, EditDriverCommand, Driver Status

+++ src/main/java/seedu/address/logic/commands/DeleteDriverCommand.java
@@ -22,6 +22,8 @@ public class DeleteDriverCommand extends Command {
             + "Example: " + COMMAND_WORD + " 1";
 
     public static final String MESSAGE_DELETE_DRIVER_SUCCESS = "Deleted Driver: %1$s";
+    public static final String MESSAGE_DELETE_DRIVER_FAIL_BUSY = "Delete Driver: %1$s "
+            + "failed because driver is not free";
 
     private final Index targetIndex;
 
@@ -39,6 +41,10 @@ public class DeleteDriverCommand extends Command {
         }
 
         Driver driverToDelete = lastShownList.get(targetIndex.getZeroBased());
+        if (driverToDelete.isBusy()) {
+            throw new CommandException(String.format(MESSAGE_DELETE_DRIVER_FAIL_BUSY, driverToDelete));
+        }
+
         model.deleteDriver(driverToDelete);
         return new CommandResult(String.format(MESSAGE_DELETE_DRIVER_SUCCESS, driverToDelete), false,
                 false, false, true, false);
